requirements.txt
requests
streamlit
python-dotenv
langchain
langgraph
pandas
numpy
nltk
pypdf
google-genai